requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
docutils==0.20.1
faiss-cpu==1.7.4
fire==0.5.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
jaraco.classes==3.2.3
keyring==23.13.1
langchain==0.0.180
markdown-it-py==2.2.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pkginfo==1.9.6
pydantic==1.10.8
Pygments==2.15.1
PyYAML==6.0
readme-renderer==37.3
regex==2023.5.5
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
six==1.16.0
SQLAlchemy==2.0.15
talk-codebase==0.1.0
tenacity==8.2.2
termcolor==2.3.0
tiktoken==0.4.0
tqdm==4.65.0
twine==4.0.2
typing-inspect==0.9.0
typing_extensions==4.6.2
urllib3==2.0.2
webencodings==0.5.1
yarl==1.9.2
zipp==3.15.0
